+++ ignite/services/chain/serve.go
@@ -461,8 +461,11 @@ func (c *Chain) start(ctx context.Context, cfg *config.ChainConfig) error {
 	// set the app as being served
 	c.served = true
 
-	// Get the first validator
-	validator := cfg.Validators[0]
+	validator, err := config.FirstValidator(cfg)
+	if err != nil {
+		return err
+	}
+
 	servers, err := validator.GetServers()
 	if err != nil {
 		return err